Add consts to LRU and use std::optional
c++17 (which our compiler supports) has std::optional<T> that can
be used for thigns that return a value or nothing (like this) Use
that.

add in some consts and pass parameters by const reference.

use std::uint32_t instead of uint. Initialize it with the constructor
initializer list instead of in the body of the constructor, and use
uniform initialization syntax to do so.

Initialize the cache instantiation using uniform initialization.

Use a few more autos so that we don't have to type const as much.